# src/main/resources/templates/js/models/reservaTablets.js
export class ReservaTablets {
    constructor(reservaCarritoTabletsId, idProfesor, idCarritoTablets, ubicacionPrestamo) {
        this.reservaCarritoTabletsId = reservaCarritoTabletsId;
        this.idProfesor = idProfesor;
        this.idCarritoTablets = idCarritoTablets;
        this.ubicacionPrestamo = ubicacionPrestamo;
    }
}